Add files via upload
Here I have used Jsonify to preview the json file of the Rest API in Python  and check whether a number is Armstrong or not . The Content-type of the Response headers is application/json which you could observe while inspecting the web preview of the app in your browser. Jsonify is a function in Flask's  json module. It serializes data to JavaScript Object Notation (JSON) format, wraps it in a Response object with the application/json mimetype. 
jsonify is useful when we are building an API in which someone would query and expect json in return.